Accept empty modules
For the Xen platform the all_video.mod module is empty.
With old binutils the .symtab section remained (containing
only section symbols), so the check didn't trigger, but starting
with binutils 2.27 not even a .symtab remains.  As there are
also no relocations that's no problem (and that is checked
independendly).