sparc64: #blocks disk node method
Return the size of the device in blocks.

Return, as an unsigned number "blocks", the number of blocks of
storage associated with the device or instance, where a "block"
is a unit of storage consisting of the number of bytes returned
by the package's "block-size" method. If the size cannot be
determined, or if the number of blocks exceeds the range of an
unsigned number, return the maximum unsigned integer (which,
because of Open Firmware's assumption of two's complement
arithmetic, is equivalent to the signed number -1).

Orabug: 24332454
Orabug: 23585494
Orabug: 23116395
Orabug: 23258099

Signed-off-by: Eric Snowberg <eric.snowberg@oracle.com>