sparc64: #blocks64 disk node method
Return the size of the device in blocks.

Return, as a double number "d.#blocks", the number of blocks of
storage associated with the device or instance, where a "block"
is a unit of storage consisting of the number of bytes returned
by the package's "block-size" method. If the size cannot be
determined, or if the number of blocks exceeds the range of a
double number, return the double number -1.

Orabug: 23257576

Signed-off-by: Eric Snowberg <eric.snowberg@oracle.com>

Reviewed-by: Thomas Tai <thomas.tai@oracle.com>