ieee1275: encode-unit command for 4 addr cell devs
Convert physical address to text unit-string.

Convert phys.lo ... phys-high, the numerical representation,
to unit-string, the text string representation of a
physical address within the address space defined by this
device node. The number of cells in the list
phys.lo ... phys.hi is determined by the value
of the #address-cells property of this node.

This function is for devices with #address-cells == 4

Orabug: 24332454
Orabug: 23585494
Orabug: 23116395
Orabug: 23258099

Signed-off-by: Eric Snowberg <eric.snowberg@oracle.com>

Reviewed-by: Thomas Tai <thomas.tai@oracle.com>